Fix defining callbacks after the method definition
This commit fixes defining callbacks after the method definition. Since
the only time #redefine_method is hit is in the #method_added hook, if a
callback was defined after the method definition the method would never
get redefined. This commit adds a call to #redefine_method in #find_or_new
so that when a callback is defined we'll check if the method has been
defined yet, and if so, redefine the method.

This commit also adds conditions to #redefine_method to short circuit if
the method hasn't been defined yet or if we've already redefined the
method by checking if the method's alias has been defined.