Fix defining callbacks before method
This commit fixes an issue when all callbacks for a given method are
defined before the method definition itself. In this case the method
wasn't being redefined to execute the callbacks since #method_added had
been renamed with a prefix. However #method_added is a hook into the
class to execute the #redefine_method when a method is added to the
class.

This commit renames #method_added and adds a couple specs testing the
scenario to make sure this doesn't break in the future.